moved ValidateBroker() to mqttbroker
